Added relevant notes + files for chapter 3
This chapter was a little hard for me. I'm still not entirely sure what
the let and where keywords do in terms of scope. I posted a thread on
the rubberduck channel with some tests, but I'm pretty sure the
conclusion I reached was not quite definitive. The understanding that I
got from the thread was that 'where' binds the definition of variables
to the expressions that call the associated variables within the do
keyword. It's a mouthful, I know. This was the conclusion I got from
learning how to and using the GHCi debugger.