Added files from exercises of chapter 6
I'm still struggling with the TypeKwonDo style exercises. Will have to
look for other examples of such and work on it more. This chapter was
slightly better than chapter 5 though, I feel like it clarifies a lot
of stuff regarding type classes and type instances which was referred
to earlier, but never really explained.