Complete chapter 7
Woohoo! Finally done with this chapter... Progress has become much
slower due to the fact that concepts have become very foreign to me, but
I am doing what I can to absorb and retain as much of this info as I
can. Regardless, I am happy that my knowledge is expanding, and hope
that I can finish the chapter 8 by the end of the week.